Revert "deleted register and added to auth"
