+++ db/users/get_user.sql
@@ -1,3 +1,3 @@
 SELECT * FROM user_yadio
 WHERE user_name = $1
- ;
+RETURNING * ;
