+++ src/index.js
@@ -5,11 +5,10 @@ import App from './App';
 import reportWebVitals from './reportWebVitals';
 import {Provider} from 'react-redux'
 import store from './redux/store'
-import {HashRouter as Router} from 'react-router-dom'
 
 ReactDOM.render(
   <React.StrictMode>
-   <Provider store={store}><Router><App /></Router> </Provider>
+   <Provider store={store}> <App /></Provider>
   </React.StrictMode>,
   document.getElementById('root')
 );
